add in slotutil
