+++ shared/slotutil/slottime.go
@@ -38,3 +38,11 @@ func EpochsSinceGenesis(genesis time.Time) types.Epoch {
 func DivideSlotBy(timesPerSlot int64) time.Duration {
 	return time.Duration(int64(params.BeaconConfig().SecondsPerSlot*1000)/timesPerSlot) * time.Millisecond
 }
+
+// AbsoluteValueSlotDifference between two slots.
+func AbsoluteValueSlotDifference(x, y types.Slot) uint64 {
+	if x > y {
+		return uint64(x.SubSlot(y))
+	}
+	return uint64(y.SubSlot(x))
+}
