+++ shared/slotutil/slottime_test.go
@@ -41,3 +41,47 @@ func TestSlotsSinceGenesis(t *testing.T) {
 		})
 	}
 }
+
+func TestAbsoluteValueSlotDifference(t *testing.T) {
+	type args struct {
+		x types.Slot
+		y types.Slot
+	}
+	tests := []struct {
+		name string
+		args args
+		want uint64
+	}{
+		{
+			name: "x_<_y",
+			args: args{
+				x: types.Slot(3),
+				y: types.Slot(4),
+			},
+			want: 1,
+		},
+		{
+			name: "x_>_y",
+			args: args{
+				x: types.Slot(100),
+				y: types.Slot(4),
+			},
+			want: 96,
+		},
+		{
+			name: "x_==_y",
+			args: args{
+				x: types.Slot(100),
+				y: types.Slot(100),
+			},
+			want: 0,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := AbsoluteValueSlotDifference(tt.args.x, tt.args.y); got != tt.want {
+				t.Errorf("AbsoluteValueSlotDifference() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
